refactor NewGrouppage to newChatPage
