+++ src/libs/Navigation/AppNavigator/ModalStackNavigators.js
@@ -2,7 +2,7 @@ import _ from 'underscore';
 import React from 'react';
 import {createStackNavigator, CardStyleInterpolators} from '@react-navigation/stack';
 import styles from '../../../styles/styles';
-import NewGroupPage from '../../../pages/NewGroupPage';
+import NewGroupPage from '../../../pages/NewChatPage';
 import SearchPage from '../../../pages/SearchPage';
 import DetailsPage from '../../../pages/DetailsPage';
 import IOURequestPage from '../../../pages/iou/IOURequestPage';
